chore: moved navbar and associated components to shared folder
