+++ commitment/imports/ui/components/shared/Layout.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Outlet } from 'react-router-dom';
-import NavBar from "../landing/NavBar";
+import NavBar from "./NavBar";
 
 
 export default function Layout() {
