+++ commitment/imports/ui/components/shared/NavBar.tsx
@@ -12,9 +12,9 @@ import {
 } from "@base/navigation-menu";
 import { useAuth } from "@hook/useAuth";
 import { Button } from "@base/button";
-import ProfileMenu from "./profile-menu";
+import {ProfileMenu} from "./profile-menu";
 
-export const NavBar: Rveact.FC = () => {
+export const NavBar: React.FC = () => {
   const { isDark, toggle } = useTheme();
 
   const location = useLocation();
